added a new directive appinfo, html template and used the new html element <app-info> in the view
In the controller MainController.js, there are three objects $scope.move, $scope.shutterbugg, and $scope.gameboard that each contain info about an app. But looking at the view (index.html), the same code is written over and over again to display each app. This is not DRY and very error-prone.

How can we solve this?
With angularJS we can teach the browser a new HTML element, for instance <app-info>, and use it in the view to display each app's details using a template instead.

First we need to make a new directive in the directive directory (js/directives), say appInfo.js. In there we use app.directive method to create the new directive named 'appInfo'. It returns an object with three options:
* restrict specifies how the directive will be used in the view. The 'E' means it will be used as a new HTML element.
* scope specifies that we will pass information into this directive through an attribute named info. The = tells the directive to look for an attribute named info in the <app-info> element, like: <app-info info="shutterbugg"></app-info>
* The data in info becomes available to use in the template given by templateURL. templateUrl specifies the HTML to use in order to display the data in scope.info. Here we use the HTML in js/directives/appInfo.html.

Secondly we create an HTML template file within our directives directory, appInfo.html. In here we define everything we would have written over and over again in our view to display details about our app: title, price, etc. Now we can use the expression info we defined in our directive to retreive the matching data from our controller.

Lastly we use the new directive as the HTML element <app-info> in our view (index.html). We pass in objects from the controller's scope ($scope.shutterbugg) into the <app-info> element's info attribute so that it displays.