created a new directive that reacts to a user's click - installapp
Directives are a core feature of AngularJS. So far we've used custom directives to simply display static content. However, directives are also a powerful way to create self-contained, interactive components. Unlike jQuery which adds interactivity as a layer on top of HTML, AngularJS treats interactivity as a native component of HTML.

How do we build this?
Let's create a new directive named 'installApp'. With restrict 'E', an empty scope, installApp.html as templateUrl, and also a fourth option link.

The link is used to create interactive directives that respond to user actions. It takes three inputs:
* scope refers to the directive's scope. Any new properties attached to $scope will become available to use in the directive's template.
* element refers to the directive's HTML element.
* attrs contains the element's attributes.

Inside the link function, there are two properties buttonText and installed, and the function download(). We'll use these in the directive's template next.

How does it work?
The template uses Angular's built-in ng-click directive. When the button is clicked, ng-click will tell AngularJS to run the download() function in the directive.

The download() function uses the scope.installed property to check if an app is installed. When an app is installed, download() does three things:
* toggles the .btn-active class
* changes the button text to "Uninstall"
* changes scope.installed to true