created new forecast service, added it to the controller and displayed it in the view
So far we've made AngularJS apps by adding data to a controller, and then displaying it in a view. But what happens when the data contains hundreds of items, or if it's constantly changing like weather or financial data? Hardcoding data into a controller won't work anymore.

A better solution is to read the live data from a server. We can do this by creating a service. Services are a way to make standalone communication logic and create independent logic that can be reused by other parts of the app.

How does it work?

You create a new service in the service directory (js/services), say forecast.js. In here we used app.factory to create a new service named forecast.
* The forecast service needs to use AngularJS's built-in $http to fetch JSON from the server. Therefore, we add $http to the forecast service as a dependency, like this: app.factory('forecast', ['$http', function($http) { ... }]);
* Then, inside forecast, we use $http to construct an HTTP GET request for the weather data. If the request succeeds, the weather data is returned; otherwise the error info is returned. Like so:
return $http.get('https://example.com/somefile.json')
	.success(function(data) { return data; })
	.error(function(err) { return err; });

Next in our controller, we want to use the forecast service to fetch data from the server. To do this:
* First add forecast as a dependency so that it's available to use. Like so:
app.controller('MainController', ['$scope', 'forecast', function($scope, forecast) { ... }]);
* Then within the controller use forecast to asynchronously fetch data from the server and store it into $scope.fiveDay. Like so: forecast.success(function(data) { $scope.fiveDay = data; });

As before, any properties attached to $scope become available to use in the view. This means in index.html, we can display the city_name using an expression as done before.