defined app routes in app.js using the app.config() mehtod and angular's $routeprovider
So far we've made AngularJS apps that display data in a single view index.html. But what happens when the app grows and needs to display more information? Stuffing more code to a single view will quickly make things messy.

A better solution is to use multiple templates that display different pieces of data based on the URL that the user is visiting. We can do this with Angular's application routes.

How do we do this?

In app.js inside the app.config() method, we use Angular's $routeProvider to define the application routes:
* We use .when() to map an URL / to a controller and template home.html. The HomeController uses the service js/services/photos.js to fetch the array of all photos from a server and stores it into $scope.photos. The home.html uses ng-repeat to loop through each item in the photos array and display each photo.
* If a user accidentally visits a URL other than /, we just redirect to / using .otherwise().

Now when a user visits /, a view will be constructed by injecting home.html into the <div ng-view></div> in index.html.